# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenSlaq API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Stack Auth JWT token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "WorkspaceWithRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Workspace ID"
          },
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "slug": {
            "type": "string",
            "description": "URL slug"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "role": {
            "type": "string",
            "description": "User's role in this workspace"
          },
          "memberCount": {
            "type": "number",
            "description": "Number of members"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "createdAt",
          "role",
          "memberCount"
        ]
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Workspace ID"
          },
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "slug": {
            "type": "string",
            "description": "URL slug"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "createdAt"
        ]
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Success indicator"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "WorkspaceInvite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Invite ID"
          },
          "workspaceId": {
            "type": "string",
            "description": "Workspace ID"
          },
          "code": {
            "type": "string",
            "description": "Invite code"
          },
          "createdBy": {
            "type": "string",
            "description": "Creator user ID"
          },
          "maxUses": {
            "type": [
              "number",
              "null"
            ],
            "description": "Maximum uses (null = unlimited)"
          },
          "useCount": {
            "type": "number",
            "description": "Number of times used"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Expiration timestamp"
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Revocation timestamp"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "code",
          "createdBy",
          "maxUses",
          "useCount",
          "expiresAt",
          "revokedAt",
          "createdAt"
        ]
      },
...
      "MarketplaceListing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "longDescription": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "requestedScopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "chat:write",
                "chat:read",
                "channels:read",
                "channels:join",
                "channels:write",
                "reactions:write",
                "reactions:read",
                "users:read",
                "presence:read",
                "channels:members:read",
                "channels:members:write",
                "commands:write"
              ]
            }
          },
          "requestedEvents": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "message:new",
                "message:updated",
                "message:deleted",
                "reaction:updated",
                "channel:updated",
                "channel:member-added",
                "channel:member-removed",
                "message:pinned",
                "message:unpinned",
                "presence:updated",
                "interaction",
                "slash_command"
              ]
            }
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "longDescription",
          "avatarUrl",
          "category",
          "requestedScopes",
          "requestedEvents",
          "published"
        ]
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tokenPrefix": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "chat:write",
                "chat:read",
                "channels:read",
                "channels:join",
                "channels:write",
                "reactions:write",
                "reactions:read",
                "users:read",
                "presence:read",
                "channels:members:read",
                "channels:members:write",
                "commands:write"
              ]
            }
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tokenPrefix",
          "scopes",
          "expiresAt",
          "lastUsedAt",
          "createdAt"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "List workspaces",
        "description": "Returns all workspaces the authenticated user is a member of.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspaces with user roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceWithRole"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Create workspace",
        "description": "Creates a new workspace and makes the authenticated user the owner.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{slug}": {
      "delete": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Delete workspace",
        "description": "Deletes a workspace. Only the workspace owner can delete.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{slug}/invites": {
      "get": {
        "tags": [
          "Invites"
        ],
        "summary": "List workspace invites",
        "description": "Returns all invites for the workspace. Requires admin permissions.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace invites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceInvite"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Create invite",
        "description": "Creates a new workspace invite link. Requires admin permissions.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxUses": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "expiresInHours": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "default": 168
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvite"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/marketplace": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "summary": "List marketplace listings",
        "description": "Returns all published marketplace bot listings.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of marketplace listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketplaceListing"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create API key",
        "description": "Creates a new user API key. Returns the full token once.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "chat:write",
                        "chat:read",
                        "channels:read",
                        "channels:join",
                        "channels:write",
                        "reactions:write",
                        "reactions:read",
                        "users:read",
                        "presence:read",
                        "channels:members:read",
                        "channels:members:write",
                        "commands:write"
                      ]
                    },
                    "minItems": 1
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "name",
                  "scopes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created API key with token",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiKey"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "token"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "description": "Lists all API keys for the current user.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiKey"
                      }
                    }
                  },
                  "required": [
                    "keys"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}